Split CRD tests
